wip remove route-action-helper
