+++ package.json
@@ -100,7 +100,6 @@
     "ember-qunit": "^5.1.5",
     "ember-resolver": "^8.0.3",
     "ember-rfc176-data": "^0.3.17",
-    "ember-route-action-helper": "^2.0.8",
     "ember-showdown-shiki": "^1.2.1",
     "ember-sinon": "^4.1.1",
     "ember-source": "~3.28.8",
